backend/requirements.txt
asgiref==3.11.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
clip @ git+https://github.com/ultralytics/CLIP.git@a6238961f9fab312d99c7709f8c7485b3f9a9b50
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==4.2.11
djangorestframework==3.15.2
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
fonttools==4.58.0
fsspec==2025.5.1
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scipy==1.15.3
seaborn==0.13.2
setuptools==80.8.0
six==1.17.0
smmap==5.0.2
sqlparse==0.5.4
sympy==1.14.0
torch==2.7.0
torchaudio==2.7.0+cpu
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.144
ultralytics-thop==2.0.14
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.11.0
